requirements.txt
git+https://github.com/Rapptz/discord.py
psycopg2-binary